+++ RELEASE_NOTES.md
@@ -13,3 +13,5 @@
 - Removed "EM" table hint for the highscore filename field in the "Script Details" section: Since not only EM tables, but also original tables use text-files for storing highscores, the "EM" hint was wrong.
 - Fixed highscore parsing for "SpongeBob's Bikini Bottom Pinball VPW". The table has a custom highscore format that was not supported yet.
 - Improved performance of recent highscores (which should fill up the corresponding dashboard widget faster).
+- Improved update button: The update button is now shown together with the target version.
+- Improved updates: The update button is now also shown when the server is older than the client, so that the update dialog can be invoked too.